mysql long progress

+++ db/mysql/src/mysql_dialect.cpp
@@ -38,7 +38,8 @@ const char* mysql_dialect::type_string(oos::data_type type) const
     case data_type::type_int:
       return "INTEGER";
     case data_type::type_long:
-      return "BIGINT";
+	  return "INTEGER";
+	  //return "BIGINT";
     case data_type::type_unsigned_char:
       return "CHAR(1)";
     case data_type::type_unsigned_short:
@@ -46,8 +47,9 @@ const char* mysql_dialect::type_string(oos::data_type type) const
     case data_type::type_unsigned_int:
       return "INTEGER";
     case data_type::type_unsigned_long:
-      return "BIGINT";
-    case data_type::type_bool:
+	  return "INTEGER";
+	  //return "BIGINT";
+	case data_type::type_bool:
       return "INTEGER";
     case data_type::type_float:
       return "FLOAT";
